修复在windows下的编译错误

+++ src/Util/SHA1.cpp
@@ -237,7 +237,7 @@ void SHA1::update(std::istream &is)
     {
         char sbuf[BLOCK_BYTES];
         is.read(sbuf, BLOCK_BYTES - buffer.size());
-        buffer.append(sbuf, is.gcount());
+        buffer.append(sbuf, static_cast<size_t>(is.gcount()));
         if (buffer.size() != BLOCK_BYTES)
         {
             return;
@@ -296,7 +296,7 @@ std::string SHA1::final_bin()
     }
 
     /* Append total_bits, split this uint64_t into two uint32_t */
-    block[BLOCK_INTS - 1] = total_bits;
+    block[BLOCK_INTS - 1] = static_cast<uint32_t>(total_bits);
     block[BLOCK_INTS - 2] = (total_bits >> 32);
     transform(digest, block, transforms);
